Assert that calculating digests precede reading content
A patch that asserts that the raw digest is calculated before the stored
(backup) digest, and that the stored digest is calculated before reading the
chunk contents.  This should ensure that any corruption to the backup due to
changing file contents would be limited to the current backup and fixed in
subsequent ones.  http://code.google.com/p/brackup/issues/detail?id=7
mentions that digests are checked on restore, so Brackup should warn if a
restored file is in an inconsistent state.  Not all asserts are added in the
case of composite chunks yet.