Specify number of threads used to restore files on command line
Patched backup-restore and Restore.pm so that the number of processes used
to restore files in parallel could be specified on the command line.
(Perhaps this should be different from the number of daemons specified for
the target as there, uploading chunks happens in parallel.) Now the code
avoids forking if the number of threads is 0 to restore the caching logic.