Patch to allow brackup to continue if files are deleted or changed
Developed a patch to allow Brackup to continue if a file disappears. I
Checking whether a chunk is in the inventory can fall back to reading from
the file if the digest cache is lost, so all of the contents of the chunk
iterator loop in Backup.pm is now wrapped in an eval.  A special prefix in
the string the code dies with indicates that it is safe to skip the current
chunk.  Another point where the code dies if a file disappears is when files
are flushed to the metafile (flush_files).  This is fixed by setting a flag
whenever errors occurred reading from $cur_file.  Modified CompositeChunk to
run StoredChunk::chunkref before modifying anything as this will die when
the underlying file changes.