Check if a daemon is already storing a chunk
Pushed a patch to check if there is a daemon already storing a given chunk,
and skip storing the current chunk in that case.  As the comparison happens
based on StoredChunk::digest, also patched the code so that StoredChunks
would use the digest of their underlying PositionedChunk.

This fixes the issue that the same PositionedChunk could be stored multiple
times in the target because the inventory is only updated once a daemon has
finished storing the chunk.  With encryption, the Stored chunks of the same
PositionedChunk also appeared to be different as encryption gives different
results each time.