New tool to re-sync the inventory with the target
Added a new command for brackup-target to re-sync the inventory with the
metafiles and chunks found on the target.  Chunks that are in a metafile but
are missing on the target or their size does not match what is stored in the
metafile are skipped and not added to the inventory.

Also, specifying --meta-dir for either gc or sync_inv allows brackup-target
to attempt to read metafiles from a local storage.  This is useful if
encryption is used, as local metafiles do not need to be decrypted.  (The
list of backups is still loaded from the target.)

The dry-run and verbose options work as expected.