Bugfixes related to sync_inv
One strange issue is that the inventory stores the full length of a
composite chunk, which is never listed in a metafile.  In this case, we now
use the length reported by the target, and make sure that it is large enough
to contain the range specified for the component chunk.