brackup-target: new fsck command that merges sync_inv and gc
Added a die to brackup-verify-inventory as it is superseded by the fsck
tool.  brackup-target: Renamed sync_inv to fsck.  Removed 'gc'.  Modified
the documentation.  Improved what fsck reports: now there are separate error
counters for each step, and most output is suppressed unless --verbose is
used.  Added garbage collection to fsck.  By default, this is skipped if
inconsistencies were found between the metafiles and the chunks on the
target, but this can be overridden using a --force-gc option.  Added logic
so that the process return value would indicate success either if there have
been no issues detected or all issues have been fixed.