Flushing and storing metafile even in case of error.
Now brackup attempts to flush and upload the metafile even if it
encounters an exception during backup. The error gets re-thrown
just before it exits. It is also noted in the metafile name if
there was something that interrupted the backup. If it was a
SIGTERM or SIGINT signal, the name ends in ".stopped.brackup"
instead of ".brackup". If it was interrupted by an exception,
the name ends in ".with-error.backup". Note: files skipped
because they could no longer be opened are not reported in this
way.