Fixed how the file digest cache is skipped.
As Root::foreach_file chdirs to the root of the source, and the default
name of the digest cache is always patched to be absolute, $path can
be relative and shorter than $digcache_file. Patched Root::foreach
to compare $digcahce_file to the completed, absolute path, and patched
the code to require the digestdb_file path to be absolute, similarly
to the path of the root of the source.