Patch to the 'accept' logic
Patched how 'accept' root configuration lines are interpreted to make sure
that parent directories of files or directories included in the backup
set are included correctly. Also fixed the issue that * was interpreted
to mean one or more characters before a non-slash as well.

Added documentation about the 'accept' feature.